requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cryptography==3.3.1
Deprecated==1.2.10
distlib==0.3.1
Django==3.1.2
django-birthday==0.1.3
django-countries==7.1
django-crispy-forms==1.9.2
django-datetime-utc==1.0.3
django-dob-widget==1.2.0
django-email-verification==0.0.7
django-phone-verify==2.0.0
django-phonenumber-field==5.0.0
djangorestframework==3.12.2
filelock==3.0.12
gunicorn==20.1.0
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
nexmo==2.5.2
numpy==1.19.2
phonenumbers==8.12.14
pickle-mixin==1.0.2
Pillow==8.0.1
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
pytesseract==0.3.6
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.1
requests==2.25.0
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
threadpoolctl==2.1.0
toml==0.10.1
twilio==6.50.0
urllib3==1.26.2
virtualenv==20.2.2
Werkzeug==1.0.1
wrapt==1.12.1
